+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -164,12 +164,14 @@ public interface DinkPluginConfig extends Config {
     )
     String tradeSection = "Player Trades";
 
+    /*
     @ConfigSection(
         name = "Leagues",
         description = "Settings for notifying when you complete league tasks, unlock areas, and redeem relics",
         position = 200,
         closedByDefault = true
     )
+     */
     String leaguesSection = "Leagues";
 
     @ConfigSection(
@@ -372,8 +374,7 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "ignoreSeasonalWorlds",
         name = "Ignore Seasonal Worlds",
-        description = "Whether to suppress notifications that occur on seasonal worlds like Leagues.<br/>" +
-            "Note: the Leagues-specific notifier uses an independent config toggle",
+        description = "Whether to suppress notifications that occur on seasonal worlds like Leagues.",
         position = 1015,
         section = advancedSection
     )
@@ -585,7 +586,8 @@ public interface DinkPluginConfig extends Config {
         description = "If non-empty, Leagues messages are sent to this URL, instead of the primary URL.<br/>" +
             "Note: this only applies to the Leagues notifier, not every notifier in a seasonal world",
         position = -1,
-        section = webhookSection
+        section = webhookSection,
+        hidden = true
     )
     default String leaguesWebhook() {
         return "";
@@ -791,7 +793,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "minLootValue",
         name = "Min Loot value",
-        description = "The minimum value of an item for a notification to be sent",
+        description = "The minimum value of an item for a notification to be sent.<br/>" +
+            "For PK chests, the <i>total</i> value of the items is compared with this threshold",
         position = 33,
         section = lootSection
     )
@@ -942,11 +945,24 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "deathIgnoredRegions",
+        name = "Ignored Region IDs",
+        description = "User-specified, comma-separated Region IDs where deaths should be ignored.<br/>" +
+            "Use the '::dinkregion' chat command or an online map to obtain the region IDs.<br/>" +
+            "For example, Prifddinas spans 12894, 12895, 13150, 13151",
+        position = 45,
+        section = deathSection
+    )
+    default String deathIgnoredRegions() {
+        return "123, 456, 789";
+    }
+
+    @ConfigItem(
         keyName = "deathMinValue",
         name = "Min Lost Value",
         description = "The minimum value of the lost items for a notification to be sent.<br/>" +
             "This setting does not apply for safe deaths",
-        position = 45,
+        position = 46,
         section = deathSection
     )
     default int deathMinValue() {
@@ -959,7 +975,7 @@ public interface DinkPluginConfig extends Config {
         description = "The message to be sent through the webhook.<br/>" +
             "Use %USERNAME% to insert your username<br/>" +
             "Use %VALUELOST% to insert the GE value of the stuff you lost",
-        position = 46,
+        position = 47,
         section = deathSection
     )
     default String deathNotifyMessage() {
@@ -970,7 +986,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "deathNotifPvpEnabled",
         name = "Distinguish PvP deaths",
         description = "Should the plugin use a different message for dying in PvP?",
-        position = 47,
+        position = 48,
         section = deathSection
     )
     default boolean deathNotifPvpEnabled() {
@@ -984,7 +1000,7 @@ public interface DinkPluginConfig extends Config {
             "Use %PKER% to insert the killer<br/>" +
             "Use %USERNAME% to insert your username<br/>" +
             "Use %VALUELOST% to insert the GE value of the stuff you lost",
-        position = 48,
+        position = 49,
         section = deathSection
     )
     default String deathNotifPvpMessage() {
@@ -1783,7 +1799,8 @@ public interface DinkPluginConfig extends Config {
         name = "Enable Leagues",
         description = "Enable notifications upon various leagues events",
         position = 200,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean notifyLeagues() {
         return false;
@@ -1794,7 +1811,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Image",
         description = "Send image with the notification",
         position = 201,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesSendImage() {
         return true;
@@ -1805,7 +1823,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Area Unlocks",
         description = "Send notifications upon area unlocks",
         position = 202,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesAreaUnlock() {
         return true;
@@ -1816,7 +1835,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Relic Unlocks",
         description = "Send notifications upon relic unlocks",
         position = 203,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesRelicUnlock() {
         return true;
@@ -1827,7 +1847,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Completed Tasks",
         description = "Send notifications upon completing a task",
         position = 204,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesTaskCompletion() {
         return true;
@@ -1838,7 +1859,8 @@ public interface DinkPluginConfig extends Config {
         name = "Task Min Difficulty",
         description = "The minimum tier of a task for a notification to be sent",
         position = 205,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default LeagueTaskDifficulty leaguesTaskMinTier() {
         return LeagueTaskDifficulty.EASY;