+++ src/main/java/dinkplugin/SettingsManager.java
@@ -9,6 +9,7 @@ import dinkplugin.notifiers.CombatTaskNotifier;
 import dinkplugin.notifiers.KillCountNotifier;
 import dinkplugin.notifiers.PetNotifier;
 import dinkplugin.util.Utils;
+import dinkplugin.util.WorldUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.Synchronized;
 import lombok.extern.slf4j.Slf4j;
@@ -16,7 +17,6 @@ import net.runelite.api.Client;
 import net.runelite.api.GameState;
 import net.runelite.api.Varbits;
 import net.runelite.api.events.CommandExecuted;
-import net.runelite.api.events.GameStateChanged;
 import net.runelite.api.events.VarbitChanged;
 import net.runelite.client.callback.ClientThread;
 import net.runelite.client.config.ConfigManager;
@@ -63,6 +63,11 @@ public class SettingsManager {
     private final Map<String, Collection<String>> keysBySection = new HashMap<>();
 
     /**
+     * The key names of config items that are hidden (and, thus, should not be exported).
+     */
+    private final Collection<String> hiddenConfigKeys = new HashSet<>();
+
+    /**
      * Set of our config keys that correspond to webhook URL lists.
      * <p>
      * These are used for special logic to merge the previous value with the new value during config imports.
@@ -102,6 +107,10 @@ public class SettingsManager {
         configManager.getConfigDescriptor(config).getItems().forEach(item -> {
             String key = item.key();
             configValueTypes.put(key, item.getType());
+
+            if (item.getItem().hidden()) {
+                hiddenConfigKeys.add(key);
+            }
 
             String section = item.getItem().section();
             if (StringUtils.isNotEmpty(section)) {
@@ -150,6 +159,9 @@ public class SettingsManager {
             }
 
             exportConfig(includeKey);
+        } else if ("DinkRegion".equalsIgnoreCase(cmd)) {
+            int regionId = WorldUtils.getLocation(client).getRegionID();
+            plugin.addChatSuccess(String.format("Your current region ID is: %d", regionId));
         }
     }
 
@@ -296,6 +308,7 @@ public class SettingsManager {
         Map<String, Object> configMap = configManager.getConfigurationKeys(prefix)
             .stream()
             .map(prop -> prop.substring(prefix.length()))
+            .filter(key -> !hiddenConfigKeys.contains(key))
             .filter(exportKey)
             .map(key -> Pair.of(key, configValueTypes.get(key)))
             .filter(pair -> pair.getValue() != null)
@@ -367,7 +380,7 @@ public class SettingsManager {
                 return;
             }
 
-            Object value = convertTypeFromJson(valueType, rawValue);
+            Object value = convertTypeFromJson(gson, valueType, rawValue);
             if (value == null) {
                 log.debug("Encountered config value with incorrect type: {} = {}", key, rawValue);
                 return;